Reorganized ComboGroups to conform to the other main combinatorial functions. Added a GetComboGroups files to ease the transition for the iterable comboGroupsIter
